requirements.txt
astroid==1.6.4
awscli==1.15.28
boto3==1.7.28
botocore==1.10.28
colorama==0.3.9
Django==1.11
django-geoposition==0.3.1
django-storages==1.6.6
djangorestframework==3.8.2
docutils==0.14
isort==4.3.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2-binary==2.7.4
pyasn1==0.4.3
pylint==1.9.1
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.4
PyYAML==3.12
raven==6.9.0
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
wrapt==1.10.11
